+++ source/extensions/lightspeed.trex.logic.ogn/python/__init__.py
@@ -1,0 +1,24 @@
+"""
+* SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+* SPDX-License-Identifier: Apache-2.0
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* https://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+"""
+
+# There is no public API to this module.
+
+__all__ = []
+
+# This import must be present in all extensions to ensure they start up and shut down properly.
+# The import is done with renaming to avoid making this internal class part of the public API.
+from ._impl.extension import PublicExtension as _PublicExtension