+++ source/extensions/lightspeed.trex.logic.ogn/python/_impl/__init__.py
@@ -1,0 +1,1 @@
+# This file can be empty. Its only reason for existing is to provide relative import capability in this directory