+++ source/extensions/lightspeed.trex.logic.widget/lightspeed/trex/logic/widget/graph_model.py
@@ -1,0 +1,25 @@
+"""
+* SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+* SPDX-License-Identifier: Apache-2.0
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* https://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+"""
+
+import carb.settings
+from omni.graph.window.core import OmniGraphModel
+
+
+class RemixLogicGraphModel(OmniGraphModel):
+    def cull_legacy_prims(self):
+        """Return True if the OgnPrim nodes should not be created in the graph"""
+        return not carb.settings.get_settings().get("/persistent/omnigraph/createPrimNodes")